+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+<!-- next-header -->
+
 ## [Unreleased](https://github.com/jeertmans/languagetool-rust/compare/v2.1.4...HEAD)
 
 ### Added
@@ -12,6 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for Mardown and Typst files.
   [#117](https://github.com/jeertmans/languagetool-rust/pull/117)
 - Added basic logger (e.g., increase verbosity with `ltrs -v`).
+  [#117](https://github.com/jeertmans/languagetool-rust/pull/117)
+
+### Changed
+
+- Changed *info* output when reading from STDIN to writing logs with INFO level,
+  instead of always writing to STDOUT.
   [#117](https://github.com/jeertmans/languagetool-rust/pull/117)
 
 ### Chore